+Dig Sign, Rem DMM ref in Inst, EA now accounted for, +ovrde notes in Inst
Added my CAC Digital Signature to prove authenticity of the document and ass macros. This will hold me personally accountable in the Gov't, and should ease any tensions n the safety of the script I write.

Made some cosmetic changes to the ProjectProperties, then made the properties password protected. Not sure if that matters, but more security the better.

Removed a referance to DMM in the Instructions, as DMM currently is removed.

Excused Absence total is now considered as Accounted For in the MasterRoster %

Added detailed notes on all the override options on the MasterRoster page in the Instructions page.