Removed Deprecated Instructions, improved CACDate efficiency
There was some old troubleshooting steps referring to a sheet that was deprecated, and another step that referred to the CACReader sheet by it's old name. This was corrected.

Removed the code I pasted into the instructions sheet and forgot to remove in the last update.

Added minor efficiency improvement to the CACDate function by removing an uneeded varriable.